% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{template}
\alias{template}
\title{Create Template}
\usage{
template(writedir = getwd())
}
\arguments{
\item{string}{directory to save Contacts.csv. Default is current directory.}
}
\description{
This function creates a template Contacts.csv file in the selected directory
(default = current directory).
The file Contacts.csv contains the project IDs and contact names
and emails that will be messaged by the script.
}
