% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{xnatNotify}
\alias{xnatNotify}
\title{XNAT Notify}
\usage{
xnatNotify(
  prearch_dir = getwd(),
  contacts = NA,
  master = NA,
  do_lastday = TRUE,
  do_lastweek = TRUE,
  do_alltime = TRUE
)
}
\arguments{
\item{prearch_dir}{prearchive directory that contains project folders.
Default is current directory.}

\item{contacts}{data.frame or string. Either a data frame containing columns
PROJECT, NAME, EMAIL, DO_FILECOUNT
or a path to a csv file containing that information. Default is to open
an interactive file selector to get the csv file.}

\item{master}{Optional master email contact. If present, they will receive
an email containing all reports sent to contacts.}
}
\description{
This function takes a prearchive location and a contact list. It searches
the prearchive for all projects found in the contact list. Then it scans
each project folder. It compiles a list of all scans there and their
file counts, and sends this information in email reports to all contacts.
An optional master email address will receive a report on all projects.
}
